requirements.txt
python-dotenv==1.0.0
openai==0.27.8
Mastodon.py==1.8.1
tweepy==4.14.0
